refactor(framework): Restructure URLs to be more logical
Currently we use a custom URL as the "overview" for a section and the
root of a section in the URL acts as the starting point, which redirects
to the first step.

This feels like it makes the URL structure less logical. With this change
removing parts of the URL now show higher levels of information.

So a section base URL will show its overview. This has introduced
the custom URL segment to signify wanting to start the user on the first
step and is only required when linking directly from the move.