chore: Attempt to gain more information from circular Type error
The production instance is currently throwing an error around the
`getComponent` with a circular structure. This error cannot be replicated
locally or any other environment.

This change attempts to provide Sentry with more information of this
error and to fail more silently, hopefully unblocking users who are
seeing this in our production environment.