build: Create Sentry releases during releases pipeline
This a set of steps to create a release, associate commits to that
releases and deploy it to an environment in Sentry.

Allowing Sentry to know about releases and deployments to environments
will allow us to track errors in much more detail as well as spot
which code changes may have caused them.