chore: Expand params debugging
This extends the debugging to try and resolve a production bug
where a circular structure is being passed to JSON.stringify.

The structure of the top level keys appears correct but somewhere in
the params object there is a circular structure.

This change is trying to show the keys for the second level objects
to try and give a clue as to where the circular structure exists.