fix: Remove circular structure when rendering components
A recent change to the way dependent fields were rendered introduced
some potential mutation of the field object.

When this field was used further down the chain, like rendering
a component it caused a circular structure to appear from the response
question.

This fix removes the chance of mutation by returning a new object.