refactor: Set user object on request
This change uses the pattern of attaching items to the request object
so that they can be used further down the request chain.

If a user is set within the session this middleware will attach that
object to the request so that it can be used elsewhere.

The main reason to do so was so that Sentry and take the user directly
from the request if one exists rather than needing to manually set the
user for Sentry.