fix(per): Correctly show banners/edit states with different move states
Currently the banners that show the status of the Person Escort Record
and allow it to be started, edited or confirmed don't have any concept
of the status of the move.

This caused issues when some moves were started or completed and the
Person Escort Record banner was no longer visible.

This tidys up the logic around when the banner exists and what users
are allowed to do in those states.