fix: Show origin on incoming dashboard
During a refactor to improve the speed of the dashboards one of the
includes needed to render the incoming dashboard was removed.

This meant that incoming moves were not grouped by their origin.

This fix adds the include for `from_location` to the API call to get
a list of moves.