fix: Allow users with PER editor role to see moves
If users only have the PER_AUTHOR role they currently can't see any
dashboards or move details because this role only adds the Person
Escort Record permissions.

This was an oversight initially and a user with this role should still
be able to view a move so that they can complete Person Escort Records.