build: Always run end-to-end tests in same instance
Previously end-to-end tests were being run against staging environments
and their respective auth services.

The staging auth service now requires two-factor authentication for some
users which prevents the tests from running as before.

For staging and PRs we moved to using a mock auth service and running
tests in the same instance on CI.

We forgot to do this for the release pipeline which caused it to fail.

This uses the same method for running tests against all environments now.