chore: Enable arbitrary items to be added to add another component
Update stored values for property if `js-property-items` set so that
reduceAddAnotherFields and reduceDependentFields generate the correct
fields for form-wizard to process the arbitrary items.

If an error ccurs when processing additional items, redirect back to
the original view to display the resulting error rather than simply
presenting the user with a 422.