chore: Handle add another values when validation errors on first-pass
Problem

Magically between the HMPO errorHandler invoking `res.redirect` and the session cookie
being written in `res.end`, the sessionModel drops the values for the add another property.

The sessionModel does not drop the errorValues for the property.

Solution

This ensures that the value for the add another property is itself unless superceded by
any value submitted or found in the sessionModels' errorValues.