chore: Remove error summary when removing add another item
When removing an item, the MOJ Add Another component regenerates
all the remaining items which results in any errors being removed
from the item. This ensures that any corresponding error summary is
also removed.

NB. when items are added, the new item is simply appended so all
errors remain intact.