feat(allocations): Move cancelled items to filter component
This moves the display of cancelled allocations to the filter component
at the top of the results.

Currently we are using a pattern of displaying cancelled results
on the same page as the active results. This means we need to make
multiple requests to the API that can become slow.

This changes that pattern to split the allocations into a separate
tab so that the results are only returned when requested.