feat(single requests): Move cancelled items to filter component
This moves the cancelled requests to a separate item in the filter
component.

Currently we display cancelled items on the same page which means
making more requests to the API. With greater volumes of data this has
caused some performance issues.

This change means that cancelled items are only loaded when requested.