refactor(per): Switch to using the batch responses endpoint
This moves answering framework responses away from the single response
endpoint to using the batch endpoint.

Currently in certain circumstances we are seeing a deadlock on the API
when creating responses asynchronously.

Using the batch endpoint allows the API to determine the order in which
it saves responses and should help resolve this issue.

It should also speed up the page speed as we only need to make a single
request to the API now.