chore: Use timer module for client metrics
Use own timer for measuring and pass this to the histogram’s observe method rather than
relying on its timer functions.

This allows for the duration to also be available to future logging capability and makes
the metrics method simpler.

Adds labels to the counters so they can be diced up by request path too.

Also adds tests for client-metrics