fix(allocation): Stop person removal from an active or completed move
Currently the API does not allow a person (or profile) to be removed
from a move that has been stared (`in_transit`) or that has taken
place(`completed`).

This change adds a check to the form wizard that handles the removal
of a person and displays a message to the user if this action cannot be
done.

Fixes BOOK-A-SECURE-MOVE-FRONTEND-K
Fixes BOOK-A-SECURE-MOVE-FRONTEND-J