chore: Capture CSRF errors in Sentry
There have been reports from users that they are seeing CSRF (This form
has been tampered with) errors frequently in production.

This change means that we will start to log them in Sentry so that we
can understand how often they are occuring and if there is a way to
reduce these occurrences.