fix: Add link to hard reload a form that has been tampered with
The content displayed on the error page when a form has been tampered
with currently instructs the user to reload the page. However, with
a number of modern browsers this will use the original request method,
in this case a `POST`, which will cause the error to continue to appear.

This change adds a link to the instruction that will force the browser
to reload this page with a `GET` which will regenerate the CSRF token
and prevent the error from re-appearing.