feat: Select radio item when only one result
This change addresses something seen from error validation analytics,
where if there is only one result when searching for a person, that
result will now be automatically selected.

The analytics showed a lot of validation errors for this particular
field, which suggests that the user behaviour is for this item to
already be selected.

Closes #809