chore: Find and inflate unpopulated resources returned to api-client
Some backend endpoints do not allow for the retrieval of resources using includes.

Devour sets any resource referenced as a relationship but not present in the
included section to null.

Modifications to `common/lib/api-client/middleware/request.js` and
`common/lib/api-client/middleware/process-response.js` allow for minimal resource stubs
to be added to the included section.

`common/lib/find-unpopulated-resources.js` allows for the detection of unpopulated resources
in any data structure passed to it

`common/lib/populate-resources.js` allows for the inflation of any unpopulated resources
in any data structure passed to it