fix(form wizard): Use consistent CSRF secret across the session
Currently each form wizard sets a CSRF token for each instance. This is
stored in the session using the SessionModel. When changes are made to
this session, for example creating a new CSRF token, the SessionModel
uses the values it had for the session at the request to save back
to the session.

This can lead to the CSRF secret being reset to a previous value in
some cases where multiple tabs are opened in quick succession.

This change overriddes the methods used to generate and verify the
CSRF token for all form wizard instances to use a globally set secret
that is stored separately on the user's session.

This will help prevent the CSRF errors that sometimes occur when a user
is interacting with different form wizards in multiple tabs.