refactor: Move framework assets to shared location
This change moves a lot of the framework specific logic (controller,
middlware, templates) used to build the Person Escort Record to a shared
location.

This change will support the re-use of the shared components to enable
the frontend to add support for other frameworks, such as the Youth
risk assessment.