refactor: Handle resource transformation in model
This ensures that each time a resource is deserialized it will be passed
through the transformer if one exists for that model.

This allows us to add any frontend specific attributes to the data
structure, for example full name within the person resource.

This method will mean it will get called regardless of what level
the model occurs in a resource's relationships.