chore: Remove tabs from move detail
This temporarily removes the tabs from the move detail view.

The API doesn't currently return the minimum events we desire so for
now the tabs won't be displayed on the detail view.

The URL will still be accessible to those who know it but it won't
be able to be accessed by navigating to it.