refactor: Remove updating of move_type
There was some code that was updating the `move_type` attribute on
the move resource.

It appeared to be doing this to pass correct context to translation
keys to be able to change content.

This seemed dangerous as other things in the future could expect
the `move_type` to be the same as what the API returns.

This change refactors that code to use the location types to determine
the correct context to set for the translation keys and keeps the
`move_type` untouched.