chore: Get framework from resource instead of setting on request
The framework is being extended to support multiple frameworks for
different resources. These resources can be created at separate points
in time and can be using different versions of the framework.

This means the framework can no longer be set once in the request
lifecycle based on the `personEscortRecord` object.

This change takes the framework from the resource and uses this when
building the relevant pieces of the UI.