refactor: Use more generic term during framework apps
Previously the only framework that we needed to support was the
Person Escort Record. Subsequently, a lot of variables were named
appropriately.

Now we need to support another type of assessment, the Youth Risk
Assessment, so these variables are being named so that they can still
apply to both (and future) use cases.