chore: Copy rejected single request rebook prop from event to move
Rebook info is not currently returned as part of the move.

Therefore we have to get the events for the move (necessitating a second call)
and fish the MoveRehject event out which does have it.

When the BE returns the rebook property either on the move directly or as
part of the meta data, this can be removed.