refactor(per): Extract banner logic to a presenter
Currently most of the logic for handling what the banner looks like
for different Person Escort Record (PER) states is done in the template
with some variables passed down to it.

This makes it a bit harder to follow and with the new need to introduce
the same banner for the youth risk assessment and its various states
this makes this more complicated.

This change extracts that logic to a presenter that passes the values
for the component straight down to the view.

This should make it easier to set the logic within the controller as
to how to structure the banner contents.