chore: Add support for different framework types in message banner
This extends the presenter responsible for showing the messsage
banner on a move.

It shows whether an assessment, for example a Person Escort Record,
can be started, edited, confirmed or printed.

This change supports the ability to set the type and pass that to the
translation stings so that they can be tailored.