fix(per): Display sections correctly in print layout
The reference to the framework was recently updated to come from the
record rather than the request object. This was to cater for multiple
assessments being supported that can use different frameworks.

The print controller was missed during that update.

This change fixes the reference to the framework for the print controller.