feat(per): Use person image in print out
Previously we have been using a placeholder section for the print out
version of the Person Escort Record.

However, we have access to an image for that person if it exists, for
example from prison.

This passes that image down to the view so that we can use that instead
of a placeholder.