fix: Apply transformer to duplicate resources correctly
In some instances, if the same resource is return multiple times within
a collection, for example a person in a list of moves, the transformer
was only being called once because devour would cache that resource within
its deserializer.

This meant that if the same person was being moved on the same day they
wouldn't have the fullname property that parts of the app expect.

This change clears the cache when an item is transformed and sets the
new values so that any future occurances return the correct structure.

This is slightly less efficient that being able to udpate a cached
resource but devour doesn't currently support a method to do this.