refactor: Remove default include from assessments
This remove the default include from the Person Escort Record and
youth risk assessments.

The default include was previously being used for all calls, including
creating and confirming an assessment which meant the API was returning
more than it needed to.

This would have caused a slighyly performance hit so the include has
been moved to only the GET calls.