refactor: Remove default include fallback in API client
When transitioning from v1 to v2 of the API we needed to define a set
of default includes to not lost functionality immediately.

Not that v2 of the API supports tailored includes we can look to only
ask for particular includes on a per request basis.

This should make requests and load times slightly faster.