test: Update devour with recent package updates
The request middleware was previously stubbed in the API client models
test.

This caused an issue where it wasn't being correctly called within the
tests and lead to the incorrect payload being returned by that
middleware.

This removes the stub and runs the middleware as expected during the
tests.

This change also includes an update to the middleware naming. Middleware
can no longer be replaced with the same name as a previous middleware.