chore: Extend assessment transformer
This extends the transformer for the assessment resource.

It includes some additions to each section of the framework to make it
easier to work with this resource within the application.

It adds:
- section progress
- responses for a particular section
- questions for a particular section
- framework question to each response

This allows some transformation that takes place within the app to
visualise responses or questions to happen a little easier.

It tries to bring together the content from the framework and the values
returned by the API into the one object.