fix: Mutate instead of return API responses when transforming
The API client library uses a mutation method when deserializing response
data and storing it into its local cache.

The initial implementation of the local transformers were built to
return a fresh set of data rather than mutate the original object as
it is often a better approach and causes less side-effects.

However, because the original library mutates and the stores nested
resources within a cache layer, when our application would return fresh
objects it would lead to incorrect data being returned when parent
objects were given a transformer.

This will now mutate the data passed to the transform methods and
also doesn't need to mess with the library's cache.

--wip-- [skip ci]