fix: Ensure next move is set dynamically for allocations
Currently we set the next unassigned move ID on load of the view
controller.

This can lead to error scenarios if a user opens mutliple tabs and
tries to go through the journey.

This change creates a new route which acts as a direct to choose the next
move that needs to be assigned. In the event that the allocation is now
full it will redirect back to the allocation.