fix: Ensure profile is checked before assigning a person to a move
This change ensures that the journey checks for an empty profile
before starting the journey to assign a person to that move.

If a person already exists this can cause issues where a person
is overriden on a move or has also caused a circular structure error
where the move is attempted to be saved to the session.

For more information see:
https://sentry.io/organizations/ministryofjustice/issues/1920836995