test: Remove cancel check from STC e2e test
This test file currently checks that a user can access the cancel link
for every scenario.

The cancellation journey is already checked within a separate test
and this set of tests is creating a new fixture on the environment
everytime which increases the test time.

Seeing as we don't have a process for cleaning up test fixtures at this
stage, this particular test is being removed.