fix: Ensure radio/checkboxes within add another keep same name
This is a fix to ensure that when items are added using JavaScript
and that item contains radios or checkboxes, that the data name does
not receive an index.

This causes each item to use a different field name and means in some
cases the form submits these values as an array instead of a string.