refactor: Use profile attribute to determine whether to show youth assessment
The logic for showing a youth risk assessment has been changed so that
the frontend now determines this.

This means that the logic for showing the banner needs to be updated to
use this new attribute rather than base it just on if the move is from
a youth establishment.