fix: Check for profile in base create controller
Currently we check if a profile exists within the assessment controller.

This is done as part of setting the previous assessment.

However, we are introducing different flows to the journey which will
mean these steps are never seen, and therefore the profile is never
created.

This leads to an error when saving the move as the profile doesn't
exist.

This also addresses a bug where if a user tries to change the person,
a new profile is not created as we don't validate that the profile is
for the currently selected person.