fix: Prevent journey error on assessment confirmation
There are a number of errors in Sentry which contain the confirmation
URL.

This appears to happen when a user finishes the confirmation journey
and hits "Back" in their browser. This lands them back on the last
step, but because the journey has been ended the form wizard kicks
in with the journey error.

This change set the `checkJourney` option to false so that it isn't
required to be checked when accessing this step.

In the case that the assessment has already been completed the other
checks on this controller kick in redirect the user back to the move.

Sentry issue: https://sentry.io/organizations/ministryofjustice/issues/2080838686/tags/url/