fix: Check move status earlier in review to prevent journey error
This change moves the position of the checks within the review controller.

If the move is not in the correct state or the user can't perform the
action, these checks will redirect to the relevant place before any
other form wizard checks are conducted, for example checking the
journey.

This prevents some of the journey errors we have been seeing which is
likely from when a user navigates back in their browser after reviewing
a move:
https://sentry.io/organizations/ministryofjustice/issues/2080838686/tags/url/