fix: Prevent stray banner when move is cancelled without PER
Currently the presenters that manage the move banners will return an
empty object if there is no assessment and the move is cancelled.

This leads to an empty banner being returned on the move detail page
instead of no banner at all.

This changes the returned value to be undefined instead.