refactor: Separate request ID and transaction ID in Sentry tags
Before we were always using the transaction ID for Sentry tags.

But in the case where the request ID doesn't exist, we don't want
to set it in Sentry, we'd rather leave it unset which shows there
was no request ID available.

This change separates the two variables and allows them to be set
independently.