fix: Prevent multiple simultaneous API token refresh calls
Currently, if the token expires on a page which requires multiple API
calls that are fired off simultaneously, it will attempt to refresh the
token for each call.

This increases the page load time and in some cases can lead to timeouts
as the API is being hit a large number of times.

This change refactors that piece of code to set a promise with the first
call. If any other subsequent calls comes in before that promise has
resolved it will wait for the first promise to resolve before returning
the token for the following calls.