chore: Omit page from sorting and filtering
This omits the page number form sorting tables or changing the filter.

If a user is viewing page 3 on a particular sort or filter, when they
change it, we need to ensure we don't persist that page number because
the number of results can change and therefore the number of pages
won't be correct.