refactor: Simplify moves redirect
The moves controllers now handle redirecting outgoing moves and single
requests to the relevant day/location when nothing is provided.

This logic was currently also replicated in the middleware for redirecting
the base moves URL.

This change simplifies that logic so that it is taken care of by
the relevant collection controller.