chore: Add support to show from or to location in the card component
This adds the ability to show the from location (destination) or
to location (pickup) in the card component when the move is displayed
on a dashboard or within an allocation.

This change is required to be able to change the grouping of moves
on the dashboard from location to vehicle.