fix: Ensure services are called with correct context
This change ensures that the service method is called on the service.

Currently it assigns the method to a variable that is called when
looking up the event type. However, when this is called the method
doesn't have the context of the initially instantiated service.

This changes the service and method to be separate properties of an
object so that the method name can be called on the original instantiated
service.

Fixes BOOK-A-SECURE-MOVE-FRONTEND-2P