chore: Expand error logging detail in Sentry
This change improves the way 422 errors are passed to Sentry.

Previously errors were sent as additional data, however, this meant that
only the first two levels of properties could be seen.

This change improves the way that the error is sent with the exception
so that more detail, like the source of meta of an error can be seen.