test: Fix unprocessable entity fixtures for suppliers
In some of the tests we create a fixture, and if no from location is
specified, the fixture helper will use the currently selected location.

However, in instances where a supplier is signed in and a random
location is selected, that location can sometimes be a court, which
won't have a supplier associated with them.

When a move is attempted to be created from these locations the API
will return a validation error.