test: Remove environment variable dependency for mock auth server
Currently the mock auth server relies on environment variables being
set to define the username and password.

This is a legacy implementation from when we used to hit the real
auth service to run end-to-end tests.

Since mocking has been introduced there is no reason to keep these
values in environment variables anymore.

This change removes the extra bit of configuration and simplifies the
implementation of the test suite.