fix: Prevent right hand panel overlapping footer content
In some instances the right hand panel will overlap the footer.

This is because the image element doesn't have a fixed height and
before the image has been fully loaded the JavaScript used to pin the
column when scrolling will set the height of the container.

When the image loads in the container is now too small so when scrolling
down or resizing the browser window, causing the column to stack,
will make the content overlap the following elements like the footer.