fix: Prevent loop when accessing the timeline
This is a temporary fix to remove the lookup for missing resources
from the frontend.

Currently the API doesn't return the full "eventable" resource within
the `included` property of the response payload. This means that the
frontend API client can't deserialize this resource.

Code was implemented to do this extra loading from the frontend and
to loop over each eventable to try and get that resource and deserialize
it.

This has caused some unexpected infinite looping in production so for
now this population is being removed and replaced with a Sentry warning
so that we can get an idea of scale of missing resources.

It might be better to implement this retrieval directly on the API too
so any major changes are being postponed until a decision on that.