fix: Use location from params on all specified routes
Param middleware can only be applied to routes that specify the param.

This sets req.location based off the session, and overrides it if specified by the location from params. This req.location is then used for the majority of the app, baring a few specific instances where the session location is used.