fix: Ensure uneditable PER message is correct for cancelled moves
This updates the content key to ensure that the correct content is
displayed for cancelled moves with an incomplete PER.

The key was previously dynamic which mean if the PER was not complete
when cancelled it wouldn't find the content.

This fixes it by hard coding that key and updating the content to
explain this further scenario.