fix: Improve missing person ID error when creating a move
We have seen a number of errors in Sentry where a profile can't
be updated because it is "missing a Person ID".

Some investigation allowed this error to be re-created when going back
to the personal details page and making a change during the move creation
journey.

This change adds some extra logic around to cater for this scenario.

It also caters for a scenario where a person is selected using the
person lookup, but then the user decides to create a new person instead.