fix: Reduce Sentry warnings for missing resources to correct instances
Currently we are throwing a sentry warning each time we try and
populate the resources for the events timeline.

In actual fact we only care about the instances when there are missing
resources.

This changes wraps this warning with a check to ensure there are
missing resources before throwing the warning to Sentry.