fix: Allow multiple moves to be created in new tabs
This fix adds support to create a new move journey in multiple tabs
using the same session.

The Form Wizard used to handle the move journeys was initially intended
for a single session purpose. We have been using it for repeat
transactions and have seen behaviour that suggests users are attempting
to create moves in multiple tabs at the same time.

This change uses the existing unique wizard configuration to create a
new wizard for each journey. The main `/new` route will now generate
a unique UUID as part of the URL so that this can be acheived.