chore: Increase logging around Form Wizard errors
This changes adds the journey details to the Sentry error when it
is triggered from within the Form Wizard.

In some instances its hard to debug what steps as user has taken
to get to that error.

This change will allow us to see the exact steps a user has accessed
during that journey. This should aid in debugging some of these errors.