refactor: Move map to a component instead of layouts and includes
This switches the map work to a separate component so that it can
be included into a view.

This changes up the previous way of loading it through a series of
layouts and includes and isolates the code a little bit more.

It also uses a separate webpack JS build just for the map work so that
it can be loaded per view rather than for all views.