fix: Don't cache failed authentication
When the auth fails, the (internal) promise is rejected without being
cleared. This means the next time the code is executed, the Promise will
automatically reject.

By clearing the Promise (and other auth token values), the next request
will have a fresh Promise to resolve or reject as per the server
response