fix: Allow handover to occur after a person has left custody
The current behavior for handing a person over is prevented once
that move has left custody. This follows previous logic tied to marking
an assessment as "confirmed".

However, handover was built to support historic recording of that
information.

This change removes the logic to prevent handover occuring once a move
has left.