fix: Ensure production assets are built and revisioned correctly
The recent Webpack upgrade caused some issues with merging
configurations, particularly with merging rules.

This meant that for production builds images and fonts were revisioned
into the root asset folder rather than the correct subfolders.

It also meant that they weren't referenced correctled in the manifest
which led to assets not loading correctly in a production environment.