fix: Ensure locations are sorted correctly
Previously locations were be sorted unix-style which is alphabetically
but case-sensitive. We have a mix of casing for location titles and
ideally need to ensure the alphabetical sorting is not case-sensitive.

This change fixes that sorting in the service closer to the source
of the data rather than at controller level.

Closes #744