chore: Add extra logging to missing Person ID error
In some instances we see an error when trying to update a profile
because it is missing a Person ID.

The current information in the Sentry errors isn't helping to pinpoint
where that data is being lost from the session.

This adds logging to capture the journey somebody has taken to get
to this point and may help understand the problem better.

It also adds extra Sentry data to errors picked up from this step
including IDs within the session and the move itself.