fix: Handle user navigating back on allocation journeys
On some of the allocation journeys, when a user navigates back they
are presented with an error saying the journey has been completed.

This should be expected behavior in some instances and we can handle
what happens when a user attempts to navigate back.

In these instances we can check for the status of the entity to prevent
the journey being accessed once this action has been completed.