fix: Prevent prereq error when user navigates back on end of journey
This adjusts the error handler in the default form wizard controller
to try and attempt to redirect the user back to the base URL of the
form journey if they have navigated back after successfully completing
a journey.

This checks the journey to see if the last visited step was the last
before save and will redirect back to the start if so.