fix: Update wording of secure youth estate triage question
This updates the wording used to determine if a person is serving
their sentence within the Secure Youth Estate or not.

We use this question to decide whether they require a youth risk
assessment and we have seen some instances where this isn't being
answered "yes" if the person is on remand.