chore: Update 422 error messages to include error title
This change updates the error message when the error is a 422 returned
by the API client.

Although an error response can contain multiple errors, in most cases
it is returning a single error.

This change updates the error message to include the title from the
first error.

The main benefit is that it will allow errors within Sentry to be
separated and potentially handled differently in terms of assigning,
ignoring or actioning.