refactor: Remove update move page object
This change removes the update move page object that was being used
for some tests within updating moves.

Each time a link to an edit step was being clicked it was creating
a new instance of the page object which didn't do anything too unique.

Instead of creating so many new instances this new approach uses
the methods that already existed in the class it was inheriting from
and removes that extra layer.