chore: Collect Sentry warnings for missing location error
We've had reports that some users are frequently seeing the missing
location 404 page. By default 404s aren't logged to Sentry with any
error information.

This change records these errors as warnings so that we can gather
some information to help debug the scenarios and understand how
frequently it is occurring.