feat: Allow getLocationsByType to accept multiple types
As of
https://github.com/ministryofjustice/hmpps-book-secure-move-api/pull/1484
the API now supports filtering on multiple types in one request.
Therefore we can update the reference data function to accept multiple
types.

I've decided not to make the change backwards compatible as there aren't
many users of the function and it's easy to change all calls to it.